Rename the files accordingly
