+++ src/sqlfluff/rules/ambiguous/AM07.py
@@ -3,7 +3,7 @@ from typing import Optional, Set, Tuple
 
 from sqlfluff.core.rules import BaseRule, LintResult, RuleContext
 from sqlfluff.core.rules.crawlers import SegmentSeekerCrawler
-from sqlfluff.utils.analysis.select_crawler import (
+from sqlfluff.utils.analysis.query import (
     Query,
     Selectable,
     WildcardInfo,
