+++ src/sqlfluff/rules/structure/ST03.py
@@ -1,7 +1,7 @@
 """Implementation of Rule ST03."""
 from sqlfluff.core.rules import BaseRule, EvalResultType, LintResult, RuleContext
 from sqlfluff.core.rules.crawlers import SegmentSeekerCrawler
-from sqlfluff.utils.analysis.select_crawler import Query
+from sqlfluff.utils.analysis.query import Query
 
 
 class Rule_ST03(BaseRule):
