+++ src/sqlfluff/utils/analysis/query.py
