+++ test/utils/analysis/test_query.py
@@ -2,7 +2,7 @@
 import pytest
 
 from sqlfluff.core.linter.linter import Linter
-from sqlfluff.utils.analysis.select_crawler import Query
+from sqlfluff.utils.analysis.query import Query
 
 
 def _parse_and_crawl_outer(sql):
